+++ src/pycaputo/integrate_fire/base.py
@@ -163,7 +163,7 @@ class IntegrateFireMethod(FractionalDifferentialEquationMethod[IntegrateFireMode
     @cached_property
     def d(self) -> tuple[CaputoDerivative, ...]:
         return tuple([
-            CaputoDerivative(order=alpha, side=Side.Left)
+            CaputoDerivative(alpha=alpha, side=Side.Left)
             for alpha in self.derivative_order
         ])